Refactor: Improve UI/UX, functionality, and animations
This commit addresses a broad request to improve the Meal Planner and Shopping List App.

The main changes include:
- Refactoring the monolithic `App.tsx` into smaller, more manageable components (`Header`, `MealPlanTab`, `RecipeTab`, `GroceryListTab`, `ChatBubble`).
- Extracting hardcoded data into separate TypeScript files (`mealData.ts`, `shoppingListData.ts`, `compliments.ts`).
- Improving the UI/UX by simplifying the header animation and enhancing the chat bubble's interactivity and design.
- Refactoring the animation implementation by moving keyframes to a global CSS file and removing inline style injection.
- Adding a `SETUP.md` file with detailed instructions on how to set up a runnable React project from the provided files.
- Replacing non-standard Figma asset imports with placeholder images to ensure the project is runnable after setup.
