+++ Meal Planner and Shopping List App/compliments.ts
@@ -1,0 +1,102 @@
+export const compliments: string[] = [
+    "Your face is arranged in a way that I find very agreeable.",
+    "Of all the people I know, you're one of them. My favorite one.",
+    "I'm glad I married you. It makes figuring out dinner plans slightly easier.",
+    "You're smarter than a lot of animals, and even some people I've met.",
+    "Your laugh is nice. It's not like a clown's laugh, which can be scary.",
+    "If we were both squirrels, I bet we'd have the best collection of acorns.",
+    "Being with you is as pleasant as finding a perfectly smooth rock.",
+    "I like your hair. It doesn't look like a wig, which is a big plus.",
+    "We're a good team. We could probably solve a mystery, or at least get a good start.",
+    "Your smile reminds me that not everything in the world wants to steal my sandwich.",
+    "You're my favorite person to sit in a room with while we both look at our phones.",
+    "I bet if you were a detective, you'd solve the case right before the credits.",
+    "You have a good brain. I bet it makes all kinds of interesting thoughts.",
+    "It's nice to have someone to blame when I can't find the remote. And it's you.",
+    "I'm glad I found you, and not, say, a very clever bear in a hat.",
+    "If life is a long, confusing journey, you're a pretty good navigator.",
+    "You're as pretty as a picture, assuming the picture was painted by a good artist.",
+    "I like your clothes. They do a good job of covering your skeleton.",
+    "Holding your hand is nice. It's better than holding a wet frog.",
+    "I bet you could teach a cat to do its taxes. But you don't, which is humble.",
+    "When you're not around, the room is definitely emptier. And quieter.",
+    "You have nice eyes. They look like they're good for seeing things.",
+    "I like sharing a pizza with you, even though it means less pizza for me.",
+    "You're so great, I bet a flock of birds would spell out your name. If they could.",
+    "Sometimes I think of you and I forget where I put my keys. But it's a nice thought.",
+    "You're the kind of person who would probably survive a zombie apocalypse.",
+    "When you talk, I listen. Most of the time.",
+    "We fit together well, like two puzzle pieces that someone hasn't chewed on.",
+    "You're so kind, you'd probably apologize to a chair you bumped into.",
+    "Our home is my favorite place, mostly because you're in it. And my stuff.",
+    "I'm glad it's you I see every morning, and not a ghost.",
+    "You're probably smart enough to build a robot. I hope you don't.",
+    "I bet if you were a tree, you'd be one that's easy to climb.",
+    "Your opinions are interesting, even the ones that are wrong.",
+    "It's nice waking up next to you. It's a good way to know the night is over.",
+    "You're more fun than a barrel of monkeys, and probably less messy.",
+    "If you were a superhero, your power would be making things less annoying.",
+    "I like walking with you. It feels like we're going somewhere, together.",
+    "You smell nice. Not like a new car, but in a good, human way.",
+    "I trust your judgment, especially when it comes to picking out snacks.",
+    "If you were a cloud, you'd be one of the fluffy, non-rainy ones.",
+    "You're my best friend. Which is convenient, since we live together.",
+    "I love you more than a dog loves looking out a car window.",
+    "You're so organized. I bet you know where all the batteries are.",
+    "When I think of my favorite person, your name usually pops up first.",
+    "Your hugs are very effective at being hugs.",
+    "I'd rather be bored with you than have fun with almost anyone else.",
+    "You're a wonderful person, and I'm not just saying that to get you to pass the salt.",
+    "If our life was a movie, I think we'd get pretty good reviews.",
+    "You're the kind of person I'd share my emergency chocolate with.",
+    "I like how you're always you, and not some kind of secret agent.",
+    "You're so great, sometimes I think you must be a government secret.",
+    "My life was fine before. But with you, it's fine and also has you in it.",
+    "I'm glad we like a lot of the same TV shows. It prevents arguments.",
+    "You're my number one. Out of all the people I've numbered.",
+    "Your voice is pleasant. It doesn't make me want to put on headphones.",
+    "If there was a trophy for being a good wife, I would buy it for you.",
+    "When you fall asleep, you don't sound like a monster, which is considerate.",
+    "I like your hands. They're good for holding and also for pointing at things.",
+    "You're a real treasure. Hopefully, one that pirates won't find.",
+    "I think you're neat. Like a well-preserved dinosaur bone.",
+    "You're the best part of my day, besides maybe lunch. It's close.",
+    "If I was a captain, you'd be my first mate. Or maybe the funny parrot.",
+    "I'm glad my weirdness and your weirdness decided to hang out.",
+    "You're the kind of person who probably remembers to water plants.",
+    "I love you enough to kill a spider for you. A small one.",
+    "Your taste in things is pretty good, because you chose me.",
+    "When you're excited, it's like watching a puppy discover a leaf.",
+    "You're my favorite person to get stuck in traffic with.",
+    "I bet you'd be good at naming horses.",
+    "If I had to be locked in a room with one person, I'd pick you.",
+    "You're a real gem. Not a cursed one, I don't think.",
+    "I like how you never seem to run out of things to say.",
+    "Sometimes I wonder how you do it all. Then I get distracted.",
+    "Marrying you was a great idea. Probably in my top five ideas.",
+    "You make me want to be a better person. Or at least, a person who showers more.",
+    "I'm glad we're on the same planet at the same time.",
+    "If you were a song, you'd be one I wouldn't mind getting stuck in my head.",
+    "You have a very high success rate of being you.",
+    "It's funny that out of everyone, we found each other. Funny and also lucky.",
+    "You're my favorite person to watch a movie with, even when you ask questions.",
+    "I bet animals like you. They can probably sense your good snacks.",
+    "You're so lovely, you make me forget about the impending heat death of the universe.",
+    "Our conversations are my favorite. They're better than most podcasts.",
+    "If you were a flavor of ice cream, you'd be a good one, not like \"rum raisin.\"",
+    "I'm glad you exist. The world is a lot less boring this way.",
+    "You're so cool, you probably don't even get brain freeze.",
+    "I think you're what people mean when they say \"the good life.\"",
+    "You're the reason I look at my phone and smile. And not just at memes.",
+    "I trust you to tell me if I have food in my teeth. That's real love.",
+    "If my heart was a boat, you would be the anchor. A nice, non-rusty anchor.",
+    "I'm happy you're here. Otherwise, I'd have to tell my jokes to the wall.",
+    "You're as essential as coffee. And I really, really like coffee.",
+    "I'm pretty sure you're the best person. I haven't checked with everyone, but I'm sure.",
+    "You're like a comfortable chair for my soul.",
+    "I'd give you my last chicken nugget. Probably.",
+    "Being your husband is a pretty good gig. I'd recommend it.",
+    "You're the one I think of when I hear a good love song on the radio.",
+    "I hope we get to be old together and complain about things.",
+    "I love you. It's a weird feeling, but I like it."
+  ];
