+++ Meal Planner and Shopping List App/mealData.ts
@@ -1,0 +1,158 @@
+import { MealData } from './types';
+
+export const mealData: MealData = {
+    "Monday": {
+      "Breakfast": {
+        name: "Greek Yogurt Power Bowl",
+        ingredients: ["1 cup plain Greek yogurt (full-fat or 2% recommended)", "1 tbsp chia seeds", "1/2 cup fresh blueberries, rinsed", "1-2 tsp honey or maple syrup"],
+        steps: "Scoop the Greek yogurt into a bowl. Sprinkle the chia seeds evenly over the top. Add the fresh blueberries. Drizzle with honey or maple syrup just before serving. Chef's Tip: For a creamier texture and to soften the chia seeds, stir them into the yogurt and let it sit for 5 minutes before adding the other toppings."
+      },
+      "Lunch": {
+        name: "Refreshing Chickpea & Herb Salad",
+        ingredients: ["1 (15-ounce) can chickpeas, rinsed and drained", "1/2 English cucumber, diced", "1 Roma tomato, diced", "1/4 red onion, finely diced", "2 tbsp extra virgin olive oil", "1 tbsp fresh lemon juice", "1/4 tsp salt", "1/8 tsp freshly ground black pepper"],
+        steps: "In a medium bowl, combine the rinsed chickpeas, diced cucumber, tomato, and red onion. In a separate small bowl or jar, whisk together the olive oil, lemon juice, salt, and pepper to create the dressing. Pour the dressing over the chickpea mixture and toss gently to combine. For best results, cover and refrigerate for at least 15 minutes to allow the flavors to marry."
+      },
+      "Dinner": {
+        name: "Sheet Pan Lemon-Herb Chicken with Roasted Vegetables & Quinoa",
+        ingredients: ["2 (6-ounce) boneless, skinless chicken breasts", "1 medium zucchini, cut into 1-inch pieces", "1 large red bell pepper, cut into 1-inch pieces", "2 tbsp olive oil", "1 tsp dried oregano", "1/2 tsp garlic powder", "Salt and freshly ground black pepper to taste", "1/2 cup uncooked quinoa", "1 cup water or broth"],
+        steps: "Prepare Quinoa: Rinse the quinoa under cold water. In a small saucepan, bring 1 cup of water or broth to a boil. Add the quinoa, reduce heat to low, cover, and simmer for 15 minutes or until the liquid is absorbed. Fluff with a fork. Prepare Chicken & Veggies: Preheat your oven to 400°F (200°C). On a large, rimmed baking sheet, place the chicken breasts and surround them with the chopped zucchini and bell pepper. Season: Drizzle everything with olive oil. Sprinkle evenly with oregano, garlic powder, salt, and pepper. Use your hands or tongs to toss everything together until well-coated. Spread into a single layer. Roast: Bake for 20-25 minutes, or until the chicken is cooked through (internal temperature reaches 165°F or 74°C) and the vegetables are tender and slightly caramelized at the edges. Serve the chicken and vegetables over a bed of the cooked quinoa."
+      },
+      "Snack": {
+        name: "Apple with Sunflower Seed Butter",
+        ingredients: ["1 apple", "2 tablespoons sunflower seed butter"],
+        steps: "Slice one apple and serve with 2 tablespoons of sunflower seed butter for dipping."
+      }
+    },
+    "Tuesday": {
+      "Breakfast": {
+        name: "Strawberry Cinnamon Overnight Oats",
+        ingredients: ["1/2 cup rolled oats", "1/2 cup oat milk (or milk of choice)", "1 tbsp ground flaxseed", "1/2 cup fresh strawberries, sliced or chopped", "1/4 tsp ground cinnamon", "Optional: 1 tsp maple syrup for sweetness"],
+        steps: "In a jar or container with a lid, combine the rolled oats, oat milk, ground flaxseed, and cinnamon. Add maple syrup if using. Stir or shake well until everything is combined. Gently stir in the sliced strawberries. Cover and refrigerate for at least 4 hours, or preferably overnight. Enjoy cold, or microwave for 60-90 seconds for a warm breakfast."
+      },
+      "Lunch": {
+        name: "Hearty Lentil Soup",
+        ingredients: ["1 tbsp olive oil", "1 small onion, finely chopped", "1 medium carrot, finely chopped", "1 celery stalk, finely chopped", "2 cloves garlic, minced", "1 cup brown or green lentils, rinsed", "4 cups vegetable or chicken broth", "1 tsp dried thyme or Italian herbs", "Salt and pepper to taste"],
+        steps: "Heat olive oil in a medium pot or Dutch oven over medium heat. Add the onion, carrot, and celery (this trio is known as a mirepoix). Sauté for 5-7 minutes, until softened. Add the minced garlic and cook for another minute until fragrant. Add the rinsed lentils, broth, and dried herbs. Bring to a boil. Reduce the heat to a simmer, cover, and cook for 25-30 minutes, or until the lentils are tender. Season generously with salt and pepper to taste."
+      },
+      "Dinner": {
+        name: "Turkey Meatballs with Marinara, Pasta & Steamed Broccoli",
+        ingredients: ["1 lb ground turkey", "1/2 cup breadcrumbs (panko or regular)", "1 large egg, lightly beaten", "2 cloves garlic, minced", "1 tsp Italian herbs", "1/2 tsp salt", "1/4 tsp black pepper", "8 oz pasta of your choice (e.g., spaghetti, penne)", "2 cups broccoli florets", "1.5 cups marinara sauce, warmed"],
+        steps: "Preheat oven to 400°F (200°C). Line a baking sheet with parchment paper. In a large bowl, combine the ground turkey, breadcrumbs, beaten egg, minced garlic, Italian herbs, salt, and pepper. Mix gently with your hands until just combined—do not overmix. Roll the mixture into 1- to 1.5-inch meatballs and place them on the prepared baking sheet. Bake for 15-20 minutes, or until cooked through and lightly browned. While meatballs bake, cook pasta according to package directions. In the last 3-4 minutes of cooking, add the broccoli florets to the pasta water to steam them. Drain the pasta and broccoli. Serve the pasta and broccoli topped with the warm marinara sauce and turkey meatballs."
+      },
+      "Snack": {
+        name: "Hummus with Veggie Sticks",
+        ingredients: ["1/4 cup hummus", "Carrot sticks", "Celery sticks"],
+        steps: "Serve 1/4 cup of hummus with a handful of carrot and celery sticks."
+      }
+    },
+    "Wednesday": {
+      "Breakfast": {
+        name: "Spinach & Feta Scrambled Eggs",
+        ingredients: ["3 large eggs", "1 tbsp milk or water", "Salt and pepper to taste", "1 tsp olive oil or butter", "1 large handful of fresh spinach (about 1-2 cups)", "2 tbsp crumbled feta cheese"],
+        steps: "In a bowl, whisk the eggs with milk or water, salt, and pepper until light and frothy. Heat oil or butter in a non-stick skillet over medium heat. Add the spinach and cook for 1-2 minutes until it wilts. Pour the egg mixture over the spinach. Cook, stirring gently with a spatula, until the eggs are cooked to your desired doneness. Remove from heat and sprinkle with crumbled feta cheese before serving."
+      },
+      "Lunch": {
+        name: "Greek Salad with White Beans",
+        ingredients: ["2 cups chopped romaine lettuce or mixed greens", "1/2 cup canned white beans (cannellini or navy), rinsed and drained", "1/2 cucumber, sliced", "1/2 cup cherry tomatoes, halved", "1/4 cup Kalamata olives", "2 tbsp crumbled feta cheese", "1 tbsp olive oil", "2 tsp lemon juice", "pinch of dried oregano"],
+        steps: "In a large bowl, combine the lettuce, white beans, cucumber, cherry tomatoes, and olives. In a small bowl, whisk together the olive oil, lemon juice, and oregano for the dressing. Pour the dressing over the salad and toss to combine. Top with crumbled feta cheese before serving."
+      },
+      "Dinner": {
+        name: "Roasted Chicken Thighs with Sweet Potatoes & Brussels Sprouts",
+        ingredients: ["4 boneless, skinless chicken thighs (about 1 lb)", "1 large sweet potato, peeled and cut into 1-inch cubes", "1.5 cups Brussels sprouts, trimmed and halved", "2 tbsp olive oil", "1 tsp paprika (smoked paprika is excellent here)", "1 tsp garlic powder", "Salt and pepper to taste"],
+        steps: "Preheat oven to 400°F (200°C). On a large, rimmed baking sheet, combine the cubed sweet potato and halved Brussels sprouts. Drizzle with 1 tbsp of olive oil and season with salt and pepper. Toss to coat and spread in an even layer. In a separate bowl, pat the chicken thighs dry. Toss them with the remaining 1 tbsp olive oil, paprika, garlic powder, salt, and pepper. Place the seasoned chicken thighs amongst the vegetables on the baking sheet. Roast for 30-35 minutes, flipping the vegetables halfway through, until the chicken is cooked through and the vegetables are tender and caramelized."
+      },
+      "Snack": {
+        name: "Grapes & Cheese",
+        ingredients: ["Handful of grapes", "Few slices of your favorite cheese (cheddar, provolone, etc.)"],
+        steps: "Serve a handful of grapes with a few slices of your favorite cheese."
+      }
+    },
+    "Thursday": {
+      "Breakfast": {
+        name: "Berry Chia Smoothie",
+        ingredients: ["1 cup oat milk or other milk", "1 ripe banana (frozen for a thicker smoothie)", "1 cup mixed berries (frozen or fresh)", "1 tbsp chia seeds"],
+        steps: "Combine all ingredients in a blender. Blend on high until completely smooth and creamy. Pour into a glass and enjoy immediately. Chef's Tip: Using frozen fruit (both banana and berries) eliminates the need for ice and creates an ultra-creamy, thick smoothie without watering down the flavor."
+      },
+      "Lunch": {
+        name: "Quinoa Tabbouleh",
+        ingredients: ["1/2 cup uncooked quinoa, cooked and cooled", "1 cup fresh parsley, finely chopped", "1/2 cup fresh mint, finely chopped", "1 Roma tomato, finely diced", "1/2 cucumber, finely diced", "2 tbsp fresh lemon juice", "2 tbsp extra virgin olive oil", "Salt and pepper to taste"],
+        steps: "Cook the quinoa according to package directions and allow it to cool completely. In a medium bowl, combine the cooled quinoa, chopped parsley, mint, tomato, and cucumber. In a small bowl, whisk together the lemon juice, olive oil, salt, and pepper. Pour the dressing over the quinoa mixture and toss gently until everything is well combined."
+      },
+      "Dinner": {
+        name: "Stuffed Bell Peppers",
+        ingredients: ["2 large bell peppers (any color), halved lengthwise and seeds removed", "1/2 lb ground turkey", "1/2 small onion, finely chopped", "2 cloves garlic, minced", "1 cup cooked rice or quinoa", "1/2 cup tomato sauce or marinara sauce", "1 tsp Italian herbs", "Salt and pepper to taste", "Optional: 1/4 cup shredded mozzarella or Parmesan cheese for topping"],
+        steps: "Preheat oven to 375°F (190°C). In a skillet over medium-high heat, cook the ground turkey with the chopped onion until the turkey is no longer pink. Add the garlic and cook for one more minute. Drain any excess fat. Remove the skillet from the heat. Stir in the cooked rice, tomato sauce, Italian herbs, salt, and pepper. Arrange the bell pepper halves in a baking dish, cut-side up. Spoon the turkey and rice mixture evenly into the pepper halves. Pour a small amount of water (about 1/4 inch) into the bottom of the baking dish to help the peppers steam. Bake for 20-25 minutes. If using cheese, sprinkle it on top and bake for another 5 minutes, or until the peppers are tender and the cheese is melted and bubbly."
+      },
+      "Snack": {
+        name: "Crispy Roasted Chickpeas",
+        ingredients: ["1 (15-ounce) can chickpeas, rinsed and drained", "1 tbsp olive oil", "1/2 tsp paprika", "1/4 tsp salt"],
+        steps: "Preheat oven to 400°F (200°C). The key to crispy chickpeas is to get them very dry. After rinsing and draining, pat them thoroughly dry with a paper towel or a clean kitchen towel. On a baking sheet, toss the dry chickpeas with olive oil, paprika, and salt until evenly coated. Roast for 20-25 minutes, shaking the pan halfway through, until they are golden brown and crispy. Let them cool slightly before eating; they will get crispier as they cool."
+      }
+    },
+    "Friday": {
+      "Breakfast": {
+        name: "Cottage Cheese with Peaches & Cinnamon",
+        ingredients: ["1 cup cottage cheese", "1 ripe peach, sliced", "1/4 tsp ground cinnamon"],
+        steps: "Spoon the cottage cheese into a bowl. Top with fresh peach slices. Sprinkle with cinnamon and serve. Chef's Tip: Choose a full-fat cottage cheese for the richest flavor and creamiest texture."
+      },
+      "Lunch": {
+        name: "Roasted Veggie & Hummus Wrap",
+        ingredients: ["1 large whole-grain tortilla or wrap", "3 tbsp hummus", "1 cup roasted vegetables (zucchini, bell peppers, onions are great)", "Handful of fresh spinach (optional)"],
+        steps: "If you need to roast vegetables: Toss chopped zucchini, peppers, and onions with olive oil, salt, and pepper. Roast at 400°F (200°C) for 20 minutes until tender. Warm the tortilla slightly in a dry pan or microwave to make it more pliable. Spread the hummus evenly over the surface of the tortilla. Layer the roasted vegetables (and spinach, if using) down the center. Fold in the sides of the tortilla and roll it up tightly."
+      },
+      "Dinner": {
+        name: "Simple Baked Eggplant Parmesan",
+        ingredients: ["1 medium eggplant, sliced into 1/2-inch thick rounds", "2 tbsp olive oil", "Salt and pepper to taste", "1 cup marinara sauce", "1 cup shredded mozzarella cheese", "Optional: Fresh basil for garnish"],
+        steps: "Preheat oven to 400°F (200°C). Line a baking sheet with parchment paper. Arrange the eggplant slices in a single layer on the baking sheet. Brush both sides with olive oil and season with salt and pepper. Bake for 15-20 minutes, flipping once, until the eggplant is tender and golden. Remove from the oven. Top each eggplant slice with a spoonful of marinara sauce and a sprinkle of mozzarella cheese. Return to the oven for 5-10 more minutes, or until the cheese is melted and bubbly. Garnish with fresh basil before serving."
+      },
+      "Snack": {
+        name: "Caprese Snack Skewers",
+        ingredients: ["Cherry tomatoes", "Fresh mozzarella balls (bocconcini)", "Extra virgin olive oil", "Fresh basil leaves"],
+        steps: "On small skewers, thread cherry tomatoes and fresh mozzarella balls (bocconcini). Drizzle with a tiny bit of extra virgin olive oil and garnish with a fresh basil leaf."
+      }
+    },
+    "Saturday": {
+      "Breakfast": {
+        name: "Yogurt with Flaxseed & Raspberries",
+        ingredients: ["1 cup plain Greek yogurt", "1 tbsp ground flaxseed", "1/2 cup fresh raspberries", "1 tsp maple syrup"],
+        steps: "In a bowl, stir the ground flaxseed directly into the Greek yogurt. Top with fresh raspberries. Drizzle with maple syrup. Chef's Tip: Stirring the ground flaxseed into the yogurt (rather than just sprinkling it on top) ensures you get its benefits in every bite and improves the overall texture."
+      },
+      "Lunch": {
+        name: "15-Minute White Bean & Kale Soup",
+        ingredients: ["1 tbsp olive oil", "1/2 onion, chopped", "2 cloves garlic, minced", "4 cups vegetable or chicken broth", "1 (15-ounce) can white beans (cannellini or navy), rinsed and drained", "2 large handfuls of kale, stems removed and leaves chopped", "Salt and pepper to taste", "Optional: Pinch of red pepper flakes"],
+        steps: "Heat olive oil in a pot over medium heat. Sauté the onion for 3-4 minutes until softened. Add the garlic and red pepper flakes (if using) and cook for another minute. Add the broth and bring to a simmer. Stir in the rinsed white beans and the chopped kale. Simmer for 10-15 minutes, or until the kale is tender. Season with salt and pepper to taste."
+      },
+      "Dinner": {
+        name: "Grilled or Baked Chicken Skewers with Brown Rice",
+        ingredients: ["1 lb boneless, skinless chicken breast, cut into 1-inch cubes", "1 green bell pepper, cut into 1-inch pieces", "1/2 red onion, cut into 1-inch wedges", "2 tbsp olive oil", "1 tsp paprika", "1/2 tsp garlic powder", "Salt and pepper to taste", "1 cup uncooked brown rice", "2 cups water or broth", "4-6 wooden or metal skewers"],
+        steps: "Prepare Rice: Cook the brown rice according to package directions. (Brown rice typically takes longer, so start this first). Prepare Skewers: If using wooden skewers, soak them in water for at least 20 minutes to prevent burning. Preheat your oven to 400°F (200°C) or preheat a grill. Season: In a large bowl, combine the chicken cubes, bell pepper, and onion. Drizzle with olive oil and sprinkle with paprika, garlic powder, salt, and pepper. Toss well to coat everything evenly. Assemble: Thread the chicken and vegetables alternately onto the skewers. Cook: Baking: Place skewers on a baking sheet lined with foil. Bake for 15-20 minutes, turning halfway through, until chicken is cooked through. Grilling: Grill over medium-high heat for 10-15 minutes, turning occasionally, until chicken is cooked and has char marks. Serve: Serve the hot skewers over a bed of brown rice."
+      },
+      "Snack": {
+        name: "Cucumber with Tzatziki",
+        ingredients: ["1 cucumber", "1/4 cup tzatziki sauce"],
+        steps: "Slice a cucumber into rounds or spears and serve with 1/4 cup of store-bought or homemade tzatziki sauce for dipping."
+      }
+    },
+    "Sunday": {
+      "Breakfast": {
+        name: "Mushroom, Spinach & Goat Cheese Omelet",
+        ingredients: ["1 tsp olive oil or butter", "1/2 cup sliced mushrooms (cremini or white)", "1 handful of fresh spinach", "3 large eggs", "1 tbsp milk or water", "Salt and pepper to taste", "1 oz (about 2 tbsp) crumbled goat cheese"],
+        steps: "Heat oil or butter in a non-stick skillet over medium heat. Add the mushrooms and sauté for 3-4 minutes until browned. Add the spinach and cook until wilted. Remove the vegetables from the pan and set aside. In a bowl, whisk the eggs with milk, salt, and pepper until frothy. Pour the egg mixture into the same skillet. As the eggs begin to set, gently push the cooked portions from the edges toward the center. When the eggs are nearly set but still soft on top, sprinkle the cooked mushrooms, spinach, and goat cheese over one half of the omelet. Carefully fold the other half over the filling. Cook for another 30 seconds, then slide onto a plate."
+      },
+      "Lunch": {
+        name: "Quick Falafel Plate",
+        ingredients: ["4-6 store-bought falafel balls", "1 cup mixed greens", "1/2 cucumber, diced", "1/2 Roma tomato, diced", "2 tbsp tahini sauce for drizzling"],
+        steps: "Heat the falafel according to package directions (this can be done in a microwave, air fryer, or oven). While the falafel heats, prepare a simple salad by placing the mixed greens, cucumber, and tomato on a plate or in a bowl. Place the warm falafel alongside the salad. Drizzle everything with the tahini sauce before serving."
+      },
+      "Dinner": {
+        name: "Deconstructed Ratatouille with Lentils & Roasted Potatoes",
+        ingredients: ["2 medium potatoes (like Yukon Gold), cut into 1-inch cubes", "1 tbsp olive oil", "Salt and pepper", "1 tbsp olive oil", "2 cloves garlic, minced", "1/2 small eggplant, diced", "1 small zucchini, diced", "1/2 red bell pepper, diced", "1 cup canned diced tomatoes (undrained)", "1 cup cooked lentils (canned or pre-cooked)", "1/2 tsp dried herbs (oregano or herbs de Provence)"],
+        steps: "Preheat oven to 400°F (200°C). On a baking sheet, toss the potato cubes with 1 tbsp olive oil, salt, and pepper. Roast for 25-30 minutes, or until golden brown and crispy, flipping halfway through. While potatoes roast, make the ratatouille. Heat 1 tbsp olive oil in a large skillet or pot over medium-high heat. Add the diced eggplant and cook for 5 minutes until it begins to soften. Add the zucchini and bell pepper and cook for another 5 minutes. Add the minced garlic and cook for 1 minute until fragrant. Stir in the diced tomatoes (with their juice), cooked lentils, and dried herbs. Bring to a simmer, then reduce heat and cook for 5-10 minutes for the flavors to meld. Season with salt and pepper to taste. Serve the lentil and vegetable ratatouille with the crispy roasted potatoes on the side or mixed in."
+      },
+      "Snack": {
+        name: "Pear with Sunflower Seed Butter",
+        ingredients: ["1 ripe pear", "2 tablespoons sunflower seed butter"],
+        steps: "Slice one ripe pear and serve with 2 tablespoons of sunflower seed butter."
+      }
+    }
+  };
